Relocate feed common tests into own file
